still not good

+++ proc.h
@@ -50,8 +50,9 @@ struct proc {
   struct inode *cwd;           // Current directory
   char name[16];               // Process name (debugging)
   int exit_status;             // exit status  
-  long long accumulator;         
-  int ps_priority;                 
+//  long long accumulator;
+  int accumulator;
+  int ps_priority;
 };
 
 // Process memory is laid out contiguously, low addresses first:
